Add skeleton of the tool
The skeleton consists of a basic python app, driver script, man page,
bash-completion script and installation script.

nest: The python app supports arguement parsing. Currently, help
(--help/-h) and version (--version) flags are included.

run-nest.py: The driver script imports the python app and is the
program run when the tool is invoked.

man: The manpage includes appropriate section headers. Dummy options
included are to be replaced with new options as and when they
are implemented.

bash-completion: Add basic auto-completion for the 'nest' command.
Currently, it supports auto-completion for the help (--help/-h)
and version (--version) flags.

install.sh: Install 'nest' command-line tool. The script should be run
as root. Running the script allows nest to be run from any directory,
adds manpage to the system and bash completion support for nest flags.

Signed-off-by: Shanthanu <shanthanu.s.rai9@gmail.com>
Signed-off-by: Dhanasekhar <sekhardhana529@gmail.com>
Signed-off-by: Narayan G <gnarayang@gmail.com>