nest: Disable default generation of config file
Config files were generated whenever nest command is run. This is not
desirable as nest command is not always run for topology generation.
Eg. "nest -h" for getting help

Hence, by default, config file generation is disabled.

The user has to make a seperate API call to generate config file at the
end of topology file.

A better solution to this is desirable.

Signed-off-by: Shanthanu <shanthanu.s.rai9@gmail.com>