nest: Ensure nest is run as root
If nest is not run as root, the user is prompted to enter root
password.

configuration.py: When nest is run was root, the config file created was
owned by root. Now, config file created is owned by user running nest.

arguements.py: "-t" option is added to nest which takes in topology
creation python file as input (Eg. any file in examples/). A seperate
process of nest is run to process and after topology is created, the
control is given back to the original nest instance.

Signed-off-by: Shanthanu <shanthanu.s.rai9@gmail.com>