topology: Fix bug in IFB implementation
IFB bandwidth was by default set to 1024 mbit and never changed to link
bandwidth. Since typically the link bandwidth is set lesser than 1024
mbit, the qdisc placed in IFB did not have any queue build up since it
was not the bottleneck.

The set_qdisc API currently sets the right bandwidth value for IFB, but
this results in it requiring an extra parameter 'min_rate', which is not
correct, but temporarily fixes the bug. A more permenant solution is
desirable.

Signed-off-by: Shanthanu <shanthanu.s.rai9@gmail.com>