cmd: Remove support for cmd
Almost all features of nest has been so far implemented as a python
API. Also, it is difficult to maintain cmd when the other modules
often keep changing.

Hence support for cmd is removed for now. Later on, it may again be
supported if cmd specific feature is to be implemented.

Signed-off-by: Shanthanu <shanthanu.s.rai9@gmail.com>