lint: Use pre-commit to enforce code standard
pre-commit is a python package that helps in installing and maintaining
pre-commit hooks in git. Using this package, the following standards
are currently enforced:

* No trailing whitespace in any line (in any file)
* End of file ended with newline
* Large files (>500KB) cannot be added in a commit.
* Pylint check is run on all staged files (sanity check)

One of the major advantages of this tool is that the developer doesn't
need to have pylint installed for running pylint tests. pre-commit
itself will download and install pylint, and run the pylint check.

Signed-off-by: Shanthanu S Rai <shanthanu.s.rai9@gmail.com>