ci: Add variable to allow pipeline run in fork
By default, we disable pipeline runs in fork repos since pipeline runs
in shared gitlab runners will fail. But at times, it is beneficial to
allow pipelines runs in fork repos, for eg., when testing some changes
in the .gitlab-ci.yml file.

A new variable FORK_PIPELINE_RUN is added in .gitlab-ci.yml, which when
set to true allows pipeline runs in fork repos (it is false by default).

Signed-off-by: Shanthanu S Rai <shanthanu.s.rai9@gmail.com>