plotter: Load seaborn-paper style from custom file
Matplotlib has recently renamed "seaborn-paper" style to
"seaborn-v0_8-paper" style. The reason for this change can be found
here:

https://matplotlib.org/stable/api/prev_api_changes/api_changes_3.6.0

Due to this, the plotter module raises error that "seaborn-paper" is
an invalid style name. To avoid such issues in the future, we have
added a custom config file that mimics "seaborn-v0_8-paper" style,
and load this file in plotter module.

Signed-off-by: Shanthanu S Rai <shanthanu.s.rai9@gmail.com>